[fix] type for SubscribeSecuDataType

+++ core/cpp/pybind/src/py_longfist.hpp
@@ -289,7 +289,7 @@ void bind(pybind11::module &&m) {
       .value("kOrderQueue", SubscribeSecuDataType::kOrderQueue)
       .export_values()
       .def("__eq__", [](const SubscribeSecuDataType &a, uint64_t b) { return static_cast<uint64_t>(a) == b; })
-      .def("__or__", py::overload_cast<const SubscribeSecuDataType &, const SubscribeSecuDataType &>(&sub_data_bitwise<SubscribeSecuDataType, int>));
+      .def("__or__", py::overload_cast<const SubscribeSecuDataType &, const SubscribeSecuDataType &>(&sub_data_bitwise<SubscribeSecuDataType, uint64_t>));
   
   py::enum_<SubscribeCategoryType>(m_enums, "SubscribeCategoryType", py::arithmetic())
       .value("kNone", SubscribeCategoryType::kNone)
